Rename feedback admin panel page
